% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_debugPrint.R
\name{debug.print}
\alias{debug.print}
\title{internal  Helper function to debug more easy}
\usage{
debug.print(
  infoLevel,
  flag = "",
  values = character(0),
  msg = "",
  type = "INFO",
  step_interval = 10,
  iter = NA
)
}
\arguments{
\item{infoLevel}{If 0, no Output (just Errors), if 1 little output, if 2 bigger output}

\item{flag}{Parse additonal info}

\item{values}{If you check variables then post this into values}

\item{msg}{here add some extra msg}

\item{type}{Type can be "INFO" or "ERROR"}

\item{step_interval}{Optional integer specifying
the interval at which debug messages are printed.}

\item{iter}{Current iteration index used for
conditional debug output.}
}
\description{
internal  Helper function to debug more easy
}
\author{
(C) 2021-2026 K. Nille-Hauf, T. Feiri, M. Ricker, T. Lux -- Hochschule Biberach (until 2022),
TU Dortmund University - Chair of Structural Concrete (since 2023)
}
\keyword{internal}
