% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logStudentT.R
\name{dlt}
\alias{dlt}
\title{Density Function for logarithmic student T distritbution}
\usage{
dlt(x, m, s, n, nue)
}
\arguments{
\item{x}{quantiles}

\item{m}{mean (1. parameter)}

\item{s}{standard deviation (2. parameter)}

\item{n}{3. paramter}

\item{nue}{degrees of freedom}
}
\value{
density
}
\description{
Density Function for logarithmic student T distritbution
}
\examples{
dlt(0.5,3,6,2,5)

}
\author{
(C) 2021 - K. Nille-Hauf, T. Feiri, M. Ricker - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau
}
