% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsModelContainer.R
\name{ensure_pkg}
\alias{ensure_pkg}
\title{Ensure that a required package is available}
\usage{
ensure_pkg(pkg)
}
\arguments{
\item{pkg}{Character string with the package name}
}
\value{
Invisible TRUE (used only for its side-effect)
}
\description{
Ensure that a required package is available
}
\note{
The function stops with an informative error if the required
  package is not installed.
}
\keyword{internal}
