% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCIS_helpers.R
\name{Internal helper functions for MC_IS()}
\alias{Internal helper functions for MC_IS()}
\title{Internal helper functions for Monte Carlo Importance Sampling (MC_IS)}
\description{
This file contains internal helper functions used by the main function
[`MC_IS()`] to perform Monte Carlo simulations with importance sampling.

The helpers defined here support:
* random number generation and cluster setup,
* creation of recorder objects for iterative logging,
* single limit state simulation routine (`MC_IS_single()`),
* conversion of recorded data to data frames after completion.

Additional helper sets will be provided for system calculations
(serial and parallel systems) in separate files.
}
\details{
These functions are not intended to be called directly by users.
They are exported internally to allow modular access within the package.
}
\seealso{
[TesiproV::MC_IS()] for the public interface that calls these helpers.
}
\author{
(C) 2021-2026 K. Nille-Hauf, T. Feiri, M. Ricker, T. Lux -- Hochschule Biberach (until 2022),
TU Dortmund University - Chair of Structural Concrete (since 2023)
}
