% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logStudentT.R
\name{LogStudentT}
\alias{LogStudentT}
\alias{dlt}
\alias{logStudentT}
\alias{lt}
\alias{plt}
\alias{qlt}
\alias{rlt}
\title{Log-Student t Distribution}
\usage{
dlt(x, hyper.param)

plt(q, hyper.param)

qlt(p, hyper.param)

rlt(n_vals, hyper.param)
}
\arguments{
\item{x, q}{Vector of quantiles.}

\item{hyper.param}{Numeric vector \eqn{(m, s, n, \nu)}.}

\item{p}{Vector of probabilities.}

\item{n_vals}{Number of random values to generate.}
}
\value{
\itemize{
  \item \code{dlt()} returns the density.
  \item \code{plt()} returns the distribution function.
  \item \code{qlt()} returns the quantile function.
  \item \code{rlt()} generates random deviates.
}
}
\description{
Density, distribution function, quantile function and random generation
for the log-Student t distribution.
}
\details{
The cumulative distribution function is given by
\deqn{
F_X(x) =
F_{t_{\nu}}\left(
\frac{\ln(X/m)}{s}
\sqrt{\frac{n}{n+1}}
\right),
}
where \eqn{F_{t_{\nu}}(\cdot)} denotes the cumulative distribution
function of a log-Student t distribution with \eqn{\nu} degrees of freedom.

The log-Student t distribution arises as the marginal
distribution of a log-normal distributed variable with conjugate
log-normal-gamma prior uncertainty in \eqn{m} and \eqn{s}.

The hyperparameter vector \code{hyper.param} is defined as
\deqn{(m, s, n, \nu)}
where
\itemize{
  \item \eqn{m} is the mean value of an equivalent sample of size \eqn{n},
  \item \eqn{s} is the empirical standard deviation of an equivalent sample
        of size \eqn{\nu + 1},
  \item \eqn{\nu} denotes the degrees of freedom of the Student t distribution.
}

For details see:
\url{https://www.jcss-lc.org/publications/jcsspmc/concrete.pdf}
}
\examples{
hp <- c(3.4, 0.14, 3, 10)
#' dlt(0.5, hp)

}
\author{
(C) 2021-2026 K. Nille-Hauf, T. Feiri, M. Ricker, T. Lux --
Hochschule Biberach (until 2022), TU Dortmund University - Chair of Structural Concrete (since 2023)
}
