% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVFOSM.R
\name{MVFOSM}
\alias{MVFOSM}
\title{Mean-Value First-Order Second-Moment (MVFOSM)}
\usage{
MVFOSM(lsf, lDistr, h = 1e-04, isExpression = FALSE, debug.level)
}
\arguments{
\item{lsf}{LSF Definition, can be Expression or Function. Defined by the FLAG isExpression (see below)}

\item{lDistr}{List of Distributions}

\item{h}{If isExpression is False, than Finite Difference Method is used for partial deviation. h is the Windowsize}

\item{isExpression}{Boolean, If TRUE lsf has to be typeof expression, otherwise lsf has to be type of function()}

\item{debug.level}{If 0 no additional info if 2 high output during calculation}
}
\value{
MVFOSM returns an object containing the following elements:
\itemize{
  \item \code{beta}: Estimated reliability index.
  \item \code{pf}: Estimated probability of failure.
  \item \code{design.point}: Design point in the original \eqn{x}-space.
  \item \code{alphas}: Direction cosines (importance factors) in the standard normal space.
  \item \code{runtime}: Total runtime of the algorithm.
}
}
\description{
Mean-Value First-Order Second-Moment (MVFOSM) method for the
approximation of failure probabilities in structural reliability analysis.

The MVFOSM method linearises the limit-state function at the mean values
of the basic random variables and estimates the reliability index
and probability of failure based on first-order moment information.

This classical approach provides a computationally efficient
approximation but may be inaccurate for strongly nonlinear
limit-state functions.
}
\references{
Freudenthal, A. M. (1956).
Safety and the probability of structural failure.
\emph{Transactions of the American Society of Civil Engineers},
121, 1337-1397.
}
\author{
(C) 2021-2026 K. Nille-Hauf, T. Feiri, M. Ricker, T. Lux -- Hochschule Biberach (until 2022),
TU Dortmund University - Chair of Structural Concrete (since 2023)
}
