% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsParam.R
\docType{class}
\name{PARAM_BASEVAR-class}
\alias{PARAM_BASEVAR-class}
\alias{PARAM_BASEVAR}
\title{Parametric Random Variable (PARAM_BASEVAR)}
\description{
Extends `PROB_BASEVAR` so that one of the distribution parameters (Mean,
Sd or DistributionType) can be **swept** over a user-defined vector of values.
}
\details{
The method `nextParam()` advances `pos` (circularly) and updates the basic
fields (`Mean`, `Sd`, `DistributionType`, `DistributionParameters`) before
calling `prepare()` again.  `getCurrentParam()` returns the *last* value that
was used (convenient for readable run names).
}
\section{Fields}{

\describe{
\item{\code{ParamValues}}{Numeric vector holding the values that will be assigned
during the sweep.}

\item{\code{ParamType}}{Which parameter is varied: `"Mean"`, `"Sd"` or
`"DistributionType"`.}

\item{\code{pos}}{Current index inside `ParamValues` (used internally).}
}}


\examples{
pvar <- PARAM_BASEVAR(
  Name = "E_mod",
  DistributionType = "norm",
  ParamType = "Mean",
  ParamValues = c(30e3, 35e3, 40e3)
)
pvar$nextParam() # sets Mean = 30.000 and prepares the distribution
pvar$getCurrentParam()

}
\author{
(C) 2021-2026 K. Nille-Hauf, T. Feiri, M. Ricker, T. Lux -- Hochschule Biberach (until 2022),
TU Dortmund University - Chair of Structural Concrete (since 2023)
}
