% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsParam.R
\docType{class}
\name{PARAM_LSF-class}
\alias{PARAM_LSF-class}
\alias{PARAM_LSF}
\title{Parametric Limit-State Function (PARAM_LSF)}
\description{
Identical to `SYS_LSF` but inherits from it to make the class hierarchy clear
when a parametric study is performed.  No additional fields or methods are
required.
}

\author{
(C) 2021-2026 K. Nille-Hauf, T. Feiri, M. Ricker, T. Lux -- Hochschule Biberach (until 2022),
TU Dortmund University - Chair of Structural Concrete (since 2023)
}
