% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logStudentT.R
\name{plt}
\alias{plt}
\title{Probablity Function for logarithmic student T distritbution}
\usage{
plt(q, m, s, n, nue)
}
\arguments{
\item{q}{quantiles}

\item{m}{mean (1. parameter)}

\item{s}{standard deviation (2. parameter)}

\item{n}{3. paramter}

\item{nue}{degrees of freedom}
}
\value{
density
}
\description{
Probablity Function for logarithmic student T distritbution
}
\author{
(C) 2021 - M. Ricker, K. Nille-Hauf, T. Feiri - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau
}
