% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsModelContainer.R
\docType{class}
\name{PROB_DETVAR-class}
\alias{PROB_DETVAR-class}
\alias{PROB_DETVAR}
\title{Object to store a deterministic parameters (PROB_DETVAR)}
\description{
A convenience subclass of `PROB_BASEVAR` that represents a deterministic
value (i.e. a variable with practically zero variance).  The object is
automatically converted to a normal distribution with an infinitesimal
standard deviation (`Sd = Mean / 1e7`).
}

\section{Inherited fields}{

\describe{
Inherits all fields of `PROB_BASEVAR` (see that documentation for a full list).
}
}

\section{Additional fields}{

\describe{
  \item{\code{Value}}{
    The deterministic value used internally as \code{Mean}.
  }
}
}

\examples{
det <- PROB_DETVAR(Name = "E_mod", Value = 30e3)
det$prepare()

}
\author{
(C) 2021-2026 K. Nille-Hauf, T. Feiri, M. Ricker, T. Lux -- Hochschule Biberach (until 2022),
TU Dortmund University - Chair of Structural Concrete (since 2023)
}
