% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logStudentT.R
\name{qlt}
\alias{qlt}
\title{Quantil Function for logarithmic student T distritbution}
\usage{
qlt(p, m, s, n, nue)
}
\arguments{
\item{p}{probablity}

\item{m}{mean (1. parameter)}

\item{s}{standard deviation (2. parameter)}

\item{n}{3. paramter}

\item{nue}{degrees of freedom}
}
\value{
quantile
}
\description{
Quantil Function for logarithmic student T distritbution
}
\author{
(C) 2021 - M. Ricker, K. Nille-Hauf, T. Feiri - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau
}
