% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logStudentT.R
\name{rlt}
\alias{rlt}
\title{Random Realisation-Function for logarithmic student T distritbution}
\usage{
rlt(n_vals, m, s, n, nue)
}
\arguments{
\item{n_vals}{number of realisations}

\item{m}{mean (1. parameter)}

\item{s}{standard deviation (2. parameter)}

\item{n}{3. paramter}

\item{nue}{degrees of freedom}
}
\value{
random number
}
\description{
Random Realisation-Function for logarithmic student T distritbution
}
\author{
(C) 2021 - M. Ricker, K. Nille-Hauf, T. Feiri - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau
}
