% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCIS_helpers.R
\name{set_cluster_rng}
\alias{set_cluster_rng}
\title{Set cluster RNG streams}
\usage{
set_cluster_rng(cl, seed)
}
\arguments{
\item{cl}{Cluster object created by `parallel::makeCluster()`.}

\item{seed}{Integer base seed used on the master process.}
}
\description{
Assigns identical RNG streams to each worker in an existing cluster.
}
\keyword{internal}
