% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_debugPrint.R
\name{debug.print}
\alias{debug.print}
\title{internal  Helper function to debug more easy}
\usage{
debug.print(infoLevel, flag = "", values, msg = "", type = "INFO")
}
\arguments{
\item{infoLevel}{If 0, no Output (just Errors), if 1 little output, if 2 bigger output}

\item{flag}{Parse additonal info}

\item{values}{If you check variables then post this into values}

\item{msg}{here add some extra msg}

\item{type}{Type can be "INFO" or "ERROR"}
}
\description{
internal  Helper function to debug more easy
}
\author{
(C) 2021 - M. Ricker, K. Nille-Hauf, T. Feiri - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau
}
