% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MVSOFM.R
\name{MVFOSM}
\alias{MVFOSM}
\title{MVFOSM}
\usage{
MVFOSM(lsf, lDistr, h = 1e-04, isExpression = FALSE, debug.level)
}
\arguments{
\item{lsf}{LSF Definition, can be Expression or Function. Defined by the FLAG isExpression (see below)}

\item{lDistr}{List of Distributions}

\item{h}{If isExpression is False, than Finite Difference Method is used for partial deviation. h is the Windowsize}

\item{isExpression}{Boolean, If TRUE lsf has to be typeof expression, otherwise lsf has to be type of function()}

\item{debug.level}{If 0 no additional info if 2 high output during calculation}
}
\value{
beta, pf, design.point in x space, alphas, runtime
}
\description{
MVFOSM
}
\references{
FREUDENTHAL, A.M. Safety and the probability of structural failure. Am Soc Civil Eng Trans 1956; 121(2843):1337–97.
}
\author{
(C) 2021 - K. Nille-Hauf, T. Feiri, M. Ricker - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau#'
}
