% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsParam.R
\docType{class}
\name{PARAM_BASEVAR-class}
\alias{PARAM_BASEVAR-class}
\alias{PARAM_BASEVAR}
\title{Object for parametric variable}
\description{
Object to create parametric basic variables
}
\section{Fields}{

\describe{
\item{\code{ParamValues}}{A vector of values of the parametric studie (e.g. c(1,3,5,7) or seq(1,10,2))}

\item{\code{ParamType}}{A field to determine what should be parametric. Possible is: "Mean", "Sd", "DistributionType"}
}}


\author{
(C) 2021 - K. Nille-Hauf, T. Feiri, M. Ricker - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau
}
