% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsParam.R
\docType{class}
\name{PARAM_DETVAR-class}
\alias{PARAM_DETVAR-class}
\alias{PARAM_DETVAR}
\title{Object for parametric deterministic variable}
\description{
Object to create parametric deterministic variables
}
\section{Fields}{

\describe{
\item{\code{ParamValues}}{A vector of values. The first element goes with the first run, second element with second run and so on.}
}}


\author{
(C) 2021 - K. Nille-Hauf, T. Feiri, M. Ricker - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau
}
