% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsParam.R
\docType{class}
\name{PARAM_LSF-class}
\alias{PARAM_LSF-class}
\alias{PARAM_LSF}
\title{System Limit State Functions}
\description{
Interface for LSF through PROB_LSF. No changes.
}

\author{
(C) 2021 - K. Nille-Hauf, T. Feiri, M. Ricker - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau
}
