% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsModelContainer.R
\docType{class}
\name{PROB_DETVAR-class}
\alias{PROB_DETVAR-class}
\alias{PROB_DETVAR}
\title{Object to store a deterministic model for base vars}
\description{
Object to store a deterministic model for base vars
}
\section{Fields}{

\describe{
\item{\code{Id}}{Place in vector of objective functional expression function(x){x[id]}}

\item{\code{Name}}{readable name like f_ck, used for transform expression to objective function}

\item{\code{Description}}{- Used for better understanding of vars}

\item{\code{Value}}{- The deterministic value that sould be used (as mean for the normal distribution with infinite small sd)}
}}


\examples{
form_rf<-PROB_MACHINE(name="FORM RF",fCall="FORM",options=list("n_optim"=20,
"loctol"=0.001, "optim_type"="rackfies"))
sorm <- PROB_MACHINE(name="SORM",fCall="SORM")
mcis<-PROB_MACHINE(name="MC IS",fCall="MC_IS",options=list("cov_user" = 0.05, "n_max"=300000))
mcsus<-PROB_MACHINE(name="MC SuS",fCall="MC_SubSam")

}
\author{
(C) 2021 - K. Nille-Hauf, T. Feiri, M. Ricker - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau
}
