% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsModelContainer.R
\docType{class}
\name{PROB_MACHINE-class}
\alias{PROB_MACHINE-class}
\alias{PROB_MACHINE}
\title{Object to store prob machines}
\description{
Object to store prob machines
}
\section{Fields}{

\describe{
\item{\code{name}}{individual name}

\item{\code{fCall}}{Function Call of the method. Possible is: "MVFOSM","FORM", "SORM", "MC_Crude", "MC_IS", "MC_SubSam"}

\item{\code{options}}{additional options for the method provided as a list. For form e.g. options=list("optim_type"="rackfies").
To get insight of all available settings of each method open the help with ?FORM, ?SORM, ?MC_IS etc.}
}}


\author{
(C) 2021 - M. Ricker, K. Nille-Hauf, T. Feiri - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau
}
