% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clsParam.R
\docType{class}
\name{SYS_PARAM-class}
\alias{SYS_PARAM-class}
\alias{SYS_PARAM}
\title{Object for parametric Studies}
\description{
Object to create probabilistic problems in parametric studies context. There are no changes how to use compared with SYS_PROB
}
\section{Fields}{

\describe{
\item{\code{beta_params}}{Outputfield: See the beta values of the studie}

\item{\code{res_params}}{Outputfield: See the the full result output of each run}
}}

\section{Methods}{

\describe{
\item{\code{printResults(path = "")}}{TesiproV can create a report file with all the necessary data for you. If you provide a path (or filename, without ending) it will store
the data there, otherwise it will report to the console. Set the path via setwd() or check it via getwd().}

\item{\code{runMachines()}}{Starts solving all given problems (sys_input) with all given algorithms (probMachines). After that one can access via $res...1}
}}

\author{
(C) 2021 - M. Ricker, K. Nille-Hauf, T. Feiri - Hochschule Biberach, Institut fuer Konstruktiven Ingenieurbau
}
